Start on abstracting formatting certain value types as Strings
This is stuck waiting for #4465. In val() I have an Object (runtime type T, or a subtype of T)
and a StringSerializer<T> that I want to invoke with it, but I can't because the stupid static typing gets in the way.

The only alternative, currently, is to use the metamodel to invoke it, but christ that's slow.